MenuController에서 지금 로그인한 사장의 가게가 맞는지 체크하는 로직 추가

+++ src/main/java/com/flab/makedel/controller/MenuController.java
@@ -5,8 +5,10 @@ import com.flab.makedel.annotation.LoginCheck;
 import com.flab.makedel.annotation.LoginCheck.UserLevel;
 import com.flab.makedel.dto.MenuDTO;
 import com.flab.makedel.service.MenuService;
+import com.flab.makedel.service.StoreService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.HttpClientErrorException;
 
 @RestController
 @RequestMapping("/stores/{storeId}/menus")
@@ -21,11 +24,17 @@ import org.springframework.web.bind.annotation.RestController;
 public class MenuController {
 
     private final MenuService menuService;
+    private final StoreService storeService;
 
     @PostMapping
     @LoginCheck(userLevel = UserLevel.OWNER)
     public void insertMenu(MenuDTO menu, @PathVariable long storeId,
         @CurrentUserId String ownerId) {
+
+        boolean isMyStore = storeService.checkMyStore(storeId, ownerId);
+        if (!isMyStore) {
+            throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
+        }
 
         MenuDTO newMenu = menuService.setStoreId(menu, storeId);
         menuService.insertMenu(newMenu);
@@ -36,6 +45,12 @@ public class MenuController {
     @LoginCheck(userLevel = UserLevel.OWNER)
     public void deleteMenu(@PathVariable Long menuId, @PathVariable long storeId,
         @CurrentUserId String ownerId) {
+
+        boolean isMyStore = storeService.checkMyStore(storeId, ownerId);
+        if (!isMyStore) {
+            throw new HttpClientErrorException(HttpStatus.UNAUTHORIZED);
+        }
+
         menuService.deleteMenu(menuId);
     }
